agent 01/requirements.txt
fastapi
uvicorn
openai
pydantic
python-dotenv
typing
requests